+++ projects/de.skuzzle.polly.plugin.rx/src/polly/rx/core/orion/model/Production.java
@@ -7,5 +7,5 @@ public interface Production extends Equatable, Comparable<Production> {
 
     public abstract RxRessource getRess();
 
-    public abstract float getRate();
+    public abstract double getRate();
 }
